chore: change metrics var types

+++ src/main/java/com/example/mcpclient/model/Metrics.java
@@ -1,12 +1,12 @@
 package com.example.mcpclient.model;
 
 public record Metrics(
-        int numberOfRowsWithMissingFields,
-        int numberOfMessagesWithBlankContent,
-        int numberOfRowsWithFieldsErrors,
-        int numberOfCallsByCountryCode,
-        int relationBetweenOkKoCalls,
-        int averageCallDurationGroupedByCountryCode,
-        int wordOcurrenceRankingInMessageContentField
+        long numberOfRowsWithMissingFields,
+        long numberOfMessagesWithBlankContent,
+        long numberOfRowsWithFieldsErrors,
+        long numberOfCallsByCountryCode,
+        long relationBetweenOkKoCalls,
+        long averageCallDurationGroupedByCountryCode,
+        long wordOcurrenceRankingInMessageContentField
 ) {
 }
